+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to the sdc11073 module will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.16] - 2022-09-29
+### Fixed
+- fixed getMetricDescriptorByCode method (translations were not handled)
+### Added
+- method have_matching_codes to compare COdedValue instances
+
 ## [1.1.15] - 2022-09-20
 ### Fixed
 - fixed missing updates in alertsByHandle, metricsByHandle.... after DescriptionsModificationReports.